added metaphlan db test

+++ resources/snakefiles/profile.smk
@@ -89,10 +89,13 @@ rule download_metaphlan_db:
         "output/benchmarks/metaphlan/download_metaphlan_db_benchmark.txt"
     shell:
         """
-            # Download the database
+        if test -f "{output}/mpa_latest"; then
+            touch {output}
+            echo "DB already installed at {output}"
+        else
             metaphlan --install --bowtie2db {output} \
-            2> {log} 1>&2
-
+                2> {log} 1>&2
+        fi
         """
